Fixes Current Tests
